Added missing update for dist_H

+++ Hyper_Geometry_Functions/dist_H.m
@@ -1,8 +1,8 @@
 
 function dist = dist_H(p1, p2)
-    dist = acosh(1 + ((real(p1) - real(p2))^2 + (imag(p1) - imag(p2))^2)/(2*imag(p1)*imag(p2)));
+    dist = acosh(1 + (double(real(p1) - real(p2))^2 + double(imag(p1) - imag(p2))^2)/(2*imag(double(p1))*imag(double(p2))));
 end
 
-% p1 = 1i;
-% p2 = 10^3 + 10^-10i;
-% dist = 2*atanh(abs((p1 - p2)/(p1 - conj(p2))))
+% dist = acosh(1 + (MCF_to_float(sum_MCF(real(p1), - real(p2)))^2 + MCF_to_float(sum_MCF(imag(p1), -imag(p2)))^2)...
+%         /...
+%         (2*MCF_to_float(imag(p1))*MCF_to_float(imag(p2))));
